// TheSharks.UI/src/types/dto/activity/ActivityListItemDto.ts
export interface ActivityListItemDto {
    id: string;
    name: string;
    date: string;
    location: string;
    locationLink: string;
    activityType: string;
    responsibleFirstName: string;
    responsibleLastName: string;
    departure: string;
    briefingTime: string;
    tide: string;
    atWater: string;
    startTime: string;
    endTime: string;
    memberCount: number;
}